RDKB-29369: Convert rbuscore from C++ to C implementation
Reason for change: We are required to disable C++ inside the rbuscore project to support vendors who wish to only use C in their builds.
Test Procedure: Need full regression testing of rbus rfc enable broadband devices.
Risks: Medium

(cherry picked from commit e162bcbe38090d5745e466490534f0a38d7068e9)

RDKB-29369: Fix coverity issues related to string copying

Reason for change: Coverity found potential overflow issue using strncpy.  Switched to strcpy with string length verification done before the copy.
Test Procedure: None
Risks: Low

(cherry picked from commit a23de01688597afee7dc01fc1d09c3cfca77fa7c)

RDKB-29369: Fix benchmark/gtestapp build failer

Reason for change: benchmark test fails to compile when enabled.  This test uses the C++ benchmark library and thus requires C++ to use.  Since this test is disabled by default and is not a required test to run, it should not be considered required to convert, thus I'm leaving the part that requires c++ as c++ in order to fix our C++ based builds which enable this test.
Test Procedure: Ensure build success if gtestapp is enable or in rbuscore.bb, -DBUILD_RBUS_SAMPLE_APPS=ON -DBUILD_RBUS_BENCHMARK_TEST=ON is added to EXTRA_OECMAKE
Risks: Low

(cherry picked from commit 673d897992edabe7666d92b1f7316224e85eb123)

RDKB-30397 : Fixed unit_test/gtestapp build failure

Reason for change: Unit test fails to compile when enabled.
This test uses the C++ gtest library and thus requires C++ to use.
Since this test is disabled by default and is not a required test to run, it should not be considered required to convert.
Thus I'm leaving the part that requires c++ as c++ in order to fix our C++ based builds which enable this test.
Test Procedure: Ensure build success when DISTRO_FEATURES_append = " gtestapp" is added
Risks: Low

(cherry picked from commit c1263de6c3e472071972f416a43009478cef111c)

+++ rtMessage_mpack.c
@@ -228,7 +228,7 @@ rtError rtMessage_ToString(rtMessage const m, char** s, uint32_t* n)
     {
         if((1 >= (size - write_offset)) || 
                 //Special handling for text as snprintf will write past a buffer boundary if precision calls for it.
-                ((MSGPACK_OBJECT_STR == m->upk.data.type) && (m->upk.data.via.str.size > (size - write_offset - 3/*account for quotes + terminator in the output*/))))
+                ((MSGPACK_OBJECT_STR == m->upk.data.type) && ((int)m->upk.data.via.str.size > (size - write_offset - 3/*account for quotes + terminator in the output*/))))
         {
             //Truncated output. Indicate so.
             //First, make room for ellipsis
